backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["takushi sato <takuchan19820801@yahoo.co.jp>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
django-rest-swagger = "^2.2.0"
django-environ = "^0.9.0"
PyJWT = "^2.6.0"
python-dateutil = "^2.8.2"
Pillow = "^9.3.0"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
